refactor: move snack provider to hooks
