+++ packages/elements/src/components/snack/snack.stories.mdx
@@ -1,8 +1,7 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks'
-import { Snack, SnackHolder } from './index'
-import { SnackProvider } from '../../contexts/snacks'
+import { SnackProvider } from '../../u/snacks'
 import { UseSnackHookStory } from './use-snack-hook.stories'
-import useSnack from '../../hooks/use-snack'
+import useSnack, { Snack, SnackHolder } from '../../hooks/use-snack'
 import { RenderHtmlMarkup } from '../../storybook/render-html-markup'
 
 <Meta title="Snack" component={Snack} />
