+++ packages/elements/src/hooks/use-snack/index.tsx
@@ -1,7 +1,36 @@
-import { useContext } from 'react'
-import { SnackContext } from '../../contexts/snacks'
-import { SnackProps } from '../../components/snack'
+import React, { createContext, useState, useRef, useContext } from 'react'
+import { SnackProps, SnackHolder } from '../../components/snack'
 
+
+export const SnackContext = createContext<SnackContextProps>({} as SnackContextProps)
+
+export const SnackProvider: React.FC = ({ children }) => {
+  const [snacks, setSnacks] = useState<SnackProps[]>([])
+
+  // must also use a ref so that inside the callback to setTimeout, `snacks` is the
+  // current value, not the value at time of the setTimeout closures invocation
+  const snackRef = useRef(snacks)
+  snackRef.current = snacks
+
+  const removeSnackById = (id: string) => {
+    setSnacks(snackRef.current.filter((snack) => snack._id !== id))
+  }
+
+  const addSnackWithTimeout = (snack: SnackProps, timeout: number) => {
+    const _id = String(Math.round(Math.random() * 100000000))
+    setSnacks([...snacks, { ...snack, _id }])
+    if (timeout > 0) {
+      window.setTimeout(() => removeSnackById(_id), timeout)
+    }
+  }
+
+  return (
+    <SnackContext.Provider value={{ addSnackWithTimeout }}>
+      <SnackHolder snacks={snacks} removeSnackById={removeSnackById} />
+      {children}
+    </SnackContext.Provider>
+  )
+}
 interface IUseSnack {
   custom: (snack: SnackProps, timeout?: number) => void
   success: (text: string, timeout?: number) => void
@@ -31,4 +60,7 @@ export default (): IUseSnack => {
   }
 
   return { custom, success, info, error, warning }
+}
+export interface SnackContextProps {
+  addSnackWithTimeout: (snack: SnackProps, timeout: number) => void
 }
